+++ packages/ui/src/lib/tabs/tab-label.ts
@@ -1,0 +1,8 @@
+import { Directive, inject, TemplateRef } from '@angular/core';
+
+@Directive({
+  selector: 'ng-template[elTabLabel]',
+})
+export class ElTabLabel {
+  readonly template = inject(TemplateRef<unknown>);
+}